+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+- `/results` route that will display the corrected address into a Google Maps view.
+
 ## Modified
 - Move `Jenkinksfile` to the root
 - Modify `Jenkinsfile` to take into account the new folder structure